requirements.txt
aws-cdk-lib==2.47.0
constructs>=10.0.0,<11.0.0
ortools==9.4.1874
numpy==1.23.4
pandas==1.5.0
boto3==1.24.95
